Lastfm: enlarge playCount field

+++ music-universe/music-data-raw-lastfm/src/main/java/yurykorzun/art/universe/music/data/raw/lastfm/api/methods/artist/topalbums/dto/ArtistTopAlbumsAlbumDto.java
@@ -15,7 +15,7 @@ import yurykorzun.art.universe.music.data.raw.lastfm.api.methods.artist.common.d
 public class ArtistTopAlbumsAlbumDto extends AlbumDto {
 
     @JsonProperty("playcount")
-    private int playCount;
+    private long playCount;
 
     @JsonProperty("artist")
     private ArtistDto artist;
