+++ music-universe/music-data-raw-lastfm/src/main/java/yurykorzun/art/universe/music/data/raw/lastfm/api/methods/artist/topalbums/processing/LastfmArtistTopAlbumsResponseProcessor.java
@@ -40,7 +40,7 @@ public class LastfmArtistTopAlbumsResponseProcessor extends LastfmApiResponsePro
     private final LastfmApiDtoProcessingService dtoProcessingService;
 
     @Value("${lastfm.client.methods.artist.topAlbums.albumPlayCountThreshold:10000}")
-    private int albumPlayCountThreshold;
+    private long albumPlayCountThreshold;
 
     private static final List<EntityAttributeHandler<LastfmAlbum, ?, ArtistTopAlbumsAlbumDto>> albumAttrHandlers;
     static {
