+++ music-universe/music-data-raw-lastfm/src/main/java/yurykorzun/art/universe/music/data/raw/lastfm/api/methods/common/dto/StatsDto.java
@@ -12,6 +12,6 @@ public class StatsDto {
     private int listeners;
 
     @JsonProperty("playcount")
-    private int playCount;
+    private long playCount;
 
 }
