+++ music-universe/music-data-raw-lastfm/src/main/java/yurykorzun/art/universe/music/data/raw/lastfm/collectable/album/entity/LastfmAlbum.java
@@ -38,10 +38,10 @@ public class LastfmAlbum extends BaseLastfmEntity {
     private String url;
 
     @Column(name = "play_count")
-    private int playCount;
+    private Long playCount;
 
     @Column(name = "listeners_count")
-    private int listenersCount;
+    private Integer listenersCount;
 
     @Column(name = "publish_ts")
     private LocalDateTime publishTs;
