+++ music-universe/music-data-raw-lastfm/src/main/java/yurykorzun/art/universe/music/data/raw/lastfm/collectable/artist/entity/LastfmArtist.java
@@ -45,7 +45,7 @@ public class LastfmArtist extends BaseLastfmEntity {
     private Integer listenersCount;
 
     @Column(name = "play_count")
-    private Integer playCount;
+    private Long playCount;
 
     @Override
     public LastfmEntityType getType() {
